refactoring aspects of the script host and the tendrils that lead into it
The overall goal at this point is to have a clean separation between the
networking (including unprocessed WebSocket messaging), the script host
(generic execution services and a system-level API), and the specific
user API for document processing.  this is to facilitate a number of
script execution modes.  the initial buildout assumed quite a lot about
what was being processed and that needs to be killed