cleaning up persistent warnings
in particular, externalizing the specific type of
WebSocketConnectionHost used to handle new connections. this should be
the last piece of the kernel that is built specifically around the
DocumentScriptEnvironment.

next step is to split the document module out.  and name it jquery?
maybe. or maybe the api.  i have no focus