reworking the events a bit
reducing coupling to guice, mainly. i want to pull the event engine out
of the typelistener and just make it a component the rest of the system
can use.  sort of leaning toward this to get rid of the use of child
injectors, which is
A. super coupled to guice
B. slow as shit

but this is just one possible route.  the scoping idea still intrigues
me as well