need a place to collect the config values
the idea is that we really just have a map that represents the values
that get set by the configuration script, and the entire set of
configuration values gets updated atomically for the sake of the rest of
the system.

code-gen to read things out might get weird.  not sure yet, javassist
sometimes gets caught up on generics