introducing http method handlers
basically, the new design of the http server is to take fuller advantage
of netty, by intelligently manipulating the pipeline with specific
handlers for whatever the situation is

moving the handling of individual request methods into different classes
is a natural

along with this is the seed of a new test suite, to help with http
conformance.  it builds a protocol test server that exercises the
complete runtime pipeline with both valid and invalid requests to ensure
that responses are as expected

this is also the first step to driving the event-driven nature of the
server out to the edge.  no more aggregating http objects blindly!  in
particular, methods that can not have a body will get upset and return
an error upon seeing one come in