the LessResource
this represents a resource used during rest processing.  it's not a very
capable resource, but it helps to hook into the dependency system for
reloading.

i might try to add a bit to the resource system to make dependencies on
creation of certain paths in pursuit of making this less system smarter