and the jasmine integration begins in earnest
this basically works, in that it correctly runs specs and reports
results.

it does not completely work, however, in that it prints results to
sysout with no way to get them via the event system. i wanted to
checkpoint the work now since, as i said, it basically works.

also I believe I would prefer the specs to be found in a different way.
but that's going to take a little thought because it's a higher-order
concern