using the broadcast api script
this necessitated a few changes - first, script initialization errors weren't being noticed, or recorded really.  just sort of muddled past.  so that is fixed

this was noticed because including API modules from scripts that aren't in the root was failing due to resolution rules, so now API modules are all prefixed with api/

and the broadcast function is being tested now so that's nice, in both unit and integration styles.