replacing the localStorage interface
this is a start.  it needs to get a little smarter, but at least now it error checks in a script compatible way

the engine package SHALL DIE. which is exciting, really. except for the jquery integration part, i feel like
that's going to take a while to get right