refactorings to support sending server events to scripts
mainly just moving some pieces into the CodeGenHelper, and creating an
in-memory classpath for javassist to hold bytes for generated classes
since apparently, despite the documentation, it doesn't do that.  yay
documentation!

on the other hand, the script event stuff actually works, but that's not
in this commit because i'm still figuring out how to test it.