optimizing resource initialization
the directory tree was requiring that every single resource receive
every single ResourceLoaded event.  that is a nightmare waiting to
explode once this gets out of being for test deployments.  extracting
the logic into a helper allows for a single listener with smarter
registration semantics with the exponential time algorithm lying in
wait.

also, the correct behavior is now asserted in the integration test, to
ensure that the expected dependency structures are maintained in the
face of such goings on.  it's fairly basic, make sure the dependencies
are what are expected at each stage of the lifecycle