introducing the ServableResource
making the notion of resources that can be served into something a little more generic. specifically, i want to kill off the "servable" structure, and just make the parts of the system the can be served understand how to serve themselves.  or at least declare their own classes for such a job.  as it is right now, it's a little too messy and not amenable to what i am aiming for.

this at least gets rid of the notion of trying to declare a server URI for things that won't get served. next step is to do the public/private split, i think