using URIMatch to find routes, so SHAs are handled uniformly
this is step one.  step two is to make extension handling part of the
trie structure. the plan is to refactor SeparatorNode logic into a base
class and make a new subclass that essentially acts the same, but with
special rules - only one per path through the trie, and it disallows
SeparatorNodes to follow it