adding a helper to expose servable info
this is the bridge to allow users to control how to route incoming
requests to particular resources.  it's also the path to killing off the
servables structure.  need to work out some details, like route
contribution (word) and the actual connection of the routes but this is
pretty much the design