fixing http client errors
the async-http-client changed its API and also does not work anymore because of a deployment error. this has led me to realize that it also brings nearly nothing to the table aside from integration headaches.  so i'm just disabling it for the moment while i do this test server bizness