optimizing continuation resumption
it was not possible to immediately restart a continuation before, as it
would trip an assertion in the continuation cache.  there was code that
tried to use this path but apparently it never got triggered.  I
accidentally found it during a weird test (which no longer really even
exists) but it's fixed now, inline resumption is OKAY